refactored structure, SJF implemented and works.
Implemented the rest of the logic to get the SJF scheduler working.
it currently uses busy waiting to see if the scheduled queue has any
awaiting requests. Something that should probably be replaced by a
pthread_cond_wait and signal interupt.

The other two schedulers (RR and MLQF) are still unimplemented.